requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
distlib==0.3.1
dj-database-url==0.5.0
Django==3.1.7
django-cleanup==5.1.0
django-heroku==0.3.1
django-liveserver==0.1a2
django-next-prev==1.1.0
easy-thumbnails==2.7.1
ffmpeg==1.4
filelock==3.0.12
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
livereload==2.6.3
moviepy==1.0.3
numpy==1.20.2
Pillow==8.1.2
proglog==0.1.9
psycopg2==2.8.6
pyfiglet==0.8.post1
pyperclip==1.8.2
python-dateutil==2.8.1
pytz==2021.1
pyuploadcare==2.7.0
requests==2.25.1
selenium==3.141.0
six==1.15.0
sorl-thumbnail==12.7.0
sqlparse==0.4.1
tornado==6.1
tqdm==4.59.0
urllib3==1.26.3
virtualenv==20.4.4
virtualenvwrapper-win==1.2.6
wget==3.2
whitenoise==5.2.0
